+++ src/main/java/com/spotinst/sdkjava/model/ISpotAdminOrganizationRepo.java
@@ -37,16 +37,7 @@ public interface ISpotAdminOrganizationRepo extends IRepository<Users, Void, Str
 
     RepoGenericResponse<List<GetAccountUserMapping>> getAccountUserMapping(String userEmail, String authToken);
 
-    RepoGenericResponse<CreateProgrammaticUserResponse> createProgrammaticUser(ProgrammaticUser createRequest,
-                                                                               String authToken) ;
-
-    RepoGenericResponse<ProgrammaticUserResponse> updateProgrammaticUser(ProgrammaticUser createRequest, String programmaticUserId, String authToken);
-
-    RepoGenericResponse<ProgrammaticUserResponse> getProgrammaticUser(String programmaticUserId, String authToken);
-
-    RepoGenericResponse<List<ProgrammaticUserResponse>> getAllProgrammaticUsers(String authToken);
-
-    RepoGenericResponse<Boolean> deleteProgrammaticUser(String programmaticUserId, String authToken);
+    RepoGenericResponse<ProgrammaticUserResponse> createProgrammaticUser(ProgrammaticUser createRequest, String authToken) ;
 
     RepoGenericResponse<CreateOrganizationResponse> createOrganization(CreateOrganization createRequest, String authToken);
