+++ src/main/java/com/spotinst/sdkjava/model/api/admin/organization/ApiProgrammaticUserResponse.java
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.spotinst.sdkjava.client.rest.IPartialUpdateEntity;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -15,11 +14,10 @@ import java.util.Set;
 @JsonFilter("PartialUpdateEntityFilter")
 public class ApiProgrammaticUserResponse implements IPartialUpdateEntity {
     @JsonIgnore
-    private Set<String>                       isSet;
-    private String                            id;
-    private String                            userName;
-    private String                            description;
-    private List<ApiProgrammaticUserAccounts> accounts;
+    private Set<String> isSet;
+    private String      token;
+    private String      name;
+    private String      id;
 
     public ApiProgrammaticUserResponse() {
         isSet = new HashSet<>();
@@ -33,6 +31,24 @@ public class ApiProgrammaticUserResponse implements IPartialUpdateEntity {
         this.isSet = isSet;
     }
 
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        isSet.add("token");
+        this.token = token;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        isSet.add("name");
+        this.name = name;
+    }
+
     public String getId() {
         return id;
     }
@@ -41,50 +57,19 @@ public class ApiProgrammaticUserResponse implements IPartialUpdateEntity {
         isSet.add("id");
         this.id = id;
     }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        isSet.add("userName");
-        this.userName = userName;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        isSet.add("description");
-        this.description = description;
-    }
-
-    public List<ApiProgrammaticUserAccounts> getAccounts() {
-        return accounts;
-    }
-
-    public void setAccounts(List<ApiProgrammaticUserAccounts> accounts) {
-        isSet.add("accounts");
-        this.accounts = accounts;
-    }
     @JsonIgnore
-    public boolean isUserNameSet() {
-        return isSet.contains("userName");
-    }
-
-    @JsonIgnore
-    public boolean isDescriptionSet() {
-        return isSet.contains("description");
-    }
-
-    @JsonIgnore
-    public boolean isAccountsSet() {
-        return isSet.contains("accounts");
+    public boolean isNameSet() {
+        return isSet.contains("name");
     }
 
     @JsonIgnore
     public boolean isIdSet() {
         return isSet.contains("id");
     }
+
+    @JsonIgnore
+    public boolean isTokenSet() {
+        return isSet.contains("token");
+    }
+
 }
