+++ src/main/java/com/spotinst/sdkjava/model/bl/admin/organization/ProgrammaticUserResponse.java
@@ -5,19 +5,16 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ProgrammaticUserResponse {
     @JsonIgnore
-    private Set<String>                    isSet;
-    private String                         id;
-    private String                         userName;
-    private String                         description;
-    private List<ProgrammaticUserAccounts> accounts;
-
+    private Set<String> isSet;
+    private String      token;
+    private String      name;
+    private String      id;
 
     public ProgrammaticUserResponse() {
         isSet = new HashSet<>();
@@ -31,6 +28,24 @@ public class ProgrammaticUserResponse {
         this.isSet = isSet;
     }
 
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        isSet.add("token");
+        this.token = token;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        isSet.add("name");
+        this.name = name;
+    }
+
     public String getId() {
         return id;
     }
@@ -40,38 +55,11 @@ public class ProgrammaticUserResponse {
         this.id = id;
     }
 
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        isSet.add("userName");
-        this.userName = userName;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        isSet.add("description");
-        this.description = description;
-    }
-
-    public List<ProgrammaticUserAccounts> getAccounts() {
-        return accounts;
-    }
-
-    public void setAccounts(List<ProgrammaticUserAccounts> accounts) {
-        isSet.add("accounts");
-        this.accounts = accounts;
-    }
-
     public static class Builder {
-        private ProgrammaticUserResponse programmaticUser;
+        private ProgrammaticUserResponse programmaticUserResponse;
 
         private Builder() {
-            this.programmaticUser = new ProgrammaticUserResponse();
+            this.programmaticUserResponse = new ProgrammaticUserResponse();
         }
 
         public static Builder get() {
@@ -80,46 +68,37 @@ public class ProgrammaticUserResponse {
         }
 
         public Builder setId(final String id) {
-            programmaticUser.setId(id);
+            programmaticUserResponse.setId(id);
             return this;
         }
 
-        public Builder setUserName(final String userName) {
-            programmaticUser.setUserName(userName);
+        public Builder setName(final String name) {
+            programmaticUserResponse.setName(name);
             return this;
         }
 
-        public Builder setDescription(final String description) {
-            programmaticUser.setDescription(description);
-            return this;
-        }
-
-        public Builder setAccounts(final List<ProgrammaticUserAccounts> accounts) {
-            programmaticUser.setAccounts(accounts);
+        public Builder setToken(final String token) {
+            programmaticUserResponse.setToken(token);
             return this;
         }
 
         public ProgrammaticUserResponse build() {
-            return programmaticUser;
+            return programmaticUserResponse;
         }
     }
-    @JsonIgnore
-    public boolean isUserNameSet() {
-        return isSet.contains("userName");
-    }
 
     @JsonIgnore
-    public boolean isDescriptionSet() {
-        return isSet.contains("description");
-    }
-
-    @JsonIgnore
-    public boolean isAccountsSet() {
-        return isSet.contains("accounts");
+    public boolean isNameSet() {
+        return isSet.contains("name");
     }
 
     @JsonIgnore
     public boolean isIdSet() {
         return isSet.contains("id");
+    }
+
+    @JsonIgnore
+    public boolean isTokenSet() {
+        return isSet.contains("token");
     }
 }
