+++ src/main/java/com/spotinst/sdkjava/model/responses/admin/organization/ProgrammaticUserApiResponse.java
@@ -1,7 +1,7 @@
 package com.spotinst.sdkjava.model.responses.admin.organization;
 
 import com.spotinst.sdkjava.client.response.BaseServiceItemsResponse;
-import com.spotinst.sdkjava.model.api.admin.organization.ApiCreateProgrammaticUserResponse;
+import com.spotinst.sdkjava.model.api.admin.organization.ApiProgrammaticUserResponse;
 
-public class ProgrammaticUserApiResponse extends BaseServiceItemsResponse<ApiCreateProgrammaticUserResponse> {
+public class ProgrammaticUserApiResponse extends BaseServiceItemsResponse<ApiProgrammaticUserResponse> {
 }
