--- src/main/java/com/spotinst/sdkjava/model/responses/admin/organization/UpdateProgrammaticUserApiResponse.java
@@ -1,7 +1,0 @@
-package com.spotinst.sdkjava.model.responses.admin.organization;
-
-import com.spotinst.sdkjava.client.response.BaseServiceItemsResponse;
-import com.spotinst.sdkjava.model.api.admin.organization.ApiProgrammaticUserResponse;
-
-public class UpdateProgrammaticUserApiResponse extends BaseServiceItemsResponse<ApiProgrammaticUserResponse> {
-}
